+++ enterprise/web-frontend/modules/baserow_enterprise/realtime.js
@@ -3,4 +3,13 @@
  * comes in, the state of the stores will be updated to match the latest update.
  */
 
-export const registerRealtimeEvents = (realtime) => {}
+export const registerRealtimeEvents = (realtime) => {
+  realtime.registerEvent(
+    'permissions_updated',
+    ({ store }, { group_id: groupId }) => {
+      if (store.getters['group/haveGroupPermissionsBeenLoaded'](groupId)) {
+        store.dispatch('notification/setPermissionsUpdated', true)
+      }
+    }
+  )
+}